Split view perspectives into individual class files
Also implemented the rest of isometric point conversion.
