+++ DropGrid.Client/DropGrid.Client.projitems
@@ -20,11 +20,14 @@
     <Compile Include="$(MSBuildThisFileDirectory)State\MenuState.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)State\GameplayState.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)State\StateId.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Map\ViewPerspectives.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Map\IMapViewPerspective.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Map\Camera.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Map\GameMapRenderer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Map\TileRenderer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Asset\AnimatedSprite.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Map\MapViewPerspectives.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Map\IsometricPerspective.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Map\CartesianPerspective.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="$(MSBuildThisFileDirectory)Content\" />
